final task done fr

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,15 +7,15 @@
  */
 int main(void)
 {
-	unsigned long long int fib1 = 1, fib2 = 2, nextTerm;
+	unsigned long int fib1 = 1, fib2 = 2, nextTerm;
 	int count;
 
-	printf("%llu, %llu", fib1, fib2);
+	printf("%lu, %lu", fib1, fib2);
 
 	for (count = 3; count <= 98; count++)
 	{
 		nextTerm = fib1 + fib2;
-		printf(", %llu", nextTerm);
+		printf(", %lu", nextTerm);
 		fib1 = fib2;
 		fib2 = nextTerm;
 	}
